完成身体数据持久化 完成提醒时间图标颜色检查

+++ MuscleRecording/MuscleRecording/Data/BodyPart.swift
@@ -151,7 +151,7 @@ enum BodyPart: String, CaseIterable, Identifiable {
 }
 
 struct Measurement {
-    var value: Double // 单位cm
+    var value: Float // 单位cm
     var range = 10...200.0
 }
